reorganised campaign explorer components into dedicated folder
