+++ scribemaster/src/components/CampaignCreationComponents/CampaignExplorerComponents/CampaignHeader.tsx
@@ -20,8 +20,8 @@ import {
   DialogHeader,
   DialogTitle,
   DialogTrigger,
-} from "../ui/dialog";
-import { Input } from "../ui/input";
+} from "../../ui/dialog";
+import { Input } from "../../ui/input";
 
 interface CampaignHeaderProps {
   campaignId: number;
