+++ scribemaster/src/components/CampaignCreationComponents/CampaignExplorerComponents/ContentViewer.tsx
