+++ scribemaster/src/pages/CampaignExplorerPage.tsx
@@ -1,8 +1,8 @@
 import { AppSidebar } from "@/components/UtilityComponents/AppSidebar";
-import CampaignExplorer from "@/components/CampaignCreationComponents/CampaignExplorer";
+import CampaignExplorer from "@/components/CampaignCreationComponents/CampaignExplorerComponents/CampaignExplorer";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import { useParams } from "react-router";
-import { CampaignExplorer2 } from "@/components/CampaignCreationComponents/CampaignExplorer2";
+import { CampaignExplorer2 } from "@/components/CampaignCreationComponents/CampaignExplorerComponents/CampaignExplorer2";
 
 const CampaignExplorerPage = () => {
   const { campaignId } = useParams<{ campaignId: string }>();
